Separate amd64-arch code from the project.
