+++ pkg/proc/core/linux_core_amd64.go
