+++ pkg/proc/stack.go
@@ -5,9 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"go/constant"
-	"strings"
 
-	"github.com/go-delve/delve/pkg/dwarf/frame"
 	"github.com/go-delve/delve/pkg/dwarf/op"
 	"github.com/go-delve/delve/pkg/dwarf/reader"
 )
@@ -268,125 +266,6 @@ func (it *stackIterator) Next() bool {
 	return true
 }
 
-// asmcgocallSPOffsetSaveSlot is the offset from systemstack.SP where
-// (goroutine.SP - StackHi) is saved in runtime.asmcgocall after the stack
-// switch happens.
-const asmcgocallSPOffsetSaveSlot = 0x28
-
-// switchStack will use the current frame to determine if it's time to
-// switch between the system stack and the goroutine stack or vice versa.
-// Sets it.atend when the top of the stack is reached.
-func (it *stackIterator) switchStack() bool {
-	if it.frame.Current.Fn == nil {
-		return false
-	}
-	switch it.frame.Current.Fn.Name {
-	case "runtime.asmcgocall":
-		if it.top || !it.systemstack {
-			return false
-		}
-
-		// This function is called by a goroutine to execute a C function and
-		// switches from the goroutine stack to the system stack.
-		// Since we are unwinding the stack from callee to caller we have  switch
-		// from the system stack to the goroutine stack.
-
-		off, _ := readIntRaw(it.mem, uintptr(it.regs.SP()+asmcgocallSPOffsetSaveSlot), int64(it.bi.Arch.PtrSize())) // reads "offset of SP from StackHi" from where runtime.asmcgocall saved it
-		oldsp := it.regs.SP()
-		it.regs.Reg(it.regs.SPRegNum).Uint64Val = uint64(int64(it.stackhi) - off)
-
-		// runtime.asmcgocall can also be called from inside the system stack,
-		// in that case no stack switch actually happens
-		if it.regs.SP() == oldsp {
-			return false
-		}
-		it.systemstack = false
-
-		// advances to the next frame in the call stack
-		it.frame.addrret = uint64(int64(it.regs.SP()) + int64(it.bi.Arch.PtrSize()))
-		it.frame.Ret, _ = readUintRaw(it.mem, uintptr(it.frame.addrret), int64(it.bi.Arch.PtrSize()))
-		it.pc = it.frame.Ret
-
-		it.top = false
-		return true
-
-	case "runtime.cgocallback_gofunc":
-		// For a detailed description of how this works read the long comment at
-		// the start of $GOROOT/src/runtime/cgocall.go and the source code of
-		// runtime.cgocallback_gofunc in $GOROOT/src/runtime/asm_amd64.s
-		//
-		// When a C functions calls back into go it will eventually call into
-		// runtime.cgocallback_gofunc which is the function that does the stack
-		// switch from the system stack back into the goroutine stack
-		// Since we are going backwards on the stack here we see the transition
-		// as goroutine stack -> system stack.
-
-		if it.top || it.systemstack {
-			return false
-		}
-
-		if it.g0_sched_sp <= 0 {
-			return false
-		}
-		// entering the system stack
-		it.regs.Reg(it.regs.SPRegNum).Uint64Val = it.g0_sched_sp
-		// reads the previous value of g0.sched.sp that runtime.cgocallback_gofunc saved on the stack
-		it.g0_sched_sp, _ = readUintRaw(it.mem, uintptr(it.regs.SP()), int64(it.bi.Arch.PtrSize()))
-		it.top = false
-		callFrameRegs, ret, retaddr := it.advanceRegs()
-		frameOnSystemStack := it.newStackframe(ret, retaddr)
-		it.pc = frameOnSystemStack.Ret
-		it.regs = callFrameRegs
-		it.systemstack = true
-		return true
-
-	case "runtime.goexit", "runtime.rt0_go", "runtime.mcall":
-		// Look for "top of stack" functions.
-		it.atend = true
-		return true
-
-	case "runtime.mstart":
-		// Calls to runtime.systemstack will switch to the systemstack then:
-		// 1. alter the goroutine stack so that it looks like systemstack_switch
-		//    was called
-		// 2. alter the system stack so that it looks like the bottom-most frame
-		//    belongs to runtime.mstart
-		// If we find a runtime.mstart frame on the system stack of a goroutine
-		// parked on runtime.systemstack_switch we assume runtime.systemstack was
-		// called and continue tracing from the parked position.
-
-		if it.top || !it.systemstack || it.g == nil {
-			return false
-		}
-		if fn := it.bi.PCToFunc(it.g.PC); fn == nil || fn.Name != "runtime.systemstack_switch" {
-			return false
-		}
-
-		it.switchToGoroutineStack()
-		return true
-
-	default:
-		if it.systemstack && it.top && it.g != nil && strings.HasPrefix(it.frame.Current.Fn.Name, "runtime.") && it.frame.Current.Fn.Name != "runtime.fatalthrow" {
-			// The runtime switches to the system stack in multiple places.
-			// This usually happens through a call to runtime.systemstack but there
-			// are functions that switch to the system stack manually (for example
-			// runtime.morestack).
-			// Since we are only interested in printing the system stack for cgo
-			// calls we switch directly to the goroutine stack if we detect that the
-			// function at the top of the stack is a runtime function.
-			//
-			// The function "runtime.fatalthrow" is deliberately excluded from this
-			// because it can end up in the stack during a cgo call and switching to
-			// the goroutine stack will exclude all the C functions from the stack
-			// trace.
-			it.switchToGoroutineStack()
-			return true
-		}
-
-		return false
-	}
-}
-
 func (it *stackIterator) switchToGoroutineStack() {
 	it.systemstack = false
 	it.top = false
@@ -417,38 +296,6 @@ func (it *stackIterator) frameBase(fn *Function) int64 {
 	}
 	fb, _, _, _ := it.bi.Location(e, dwarf.AttrFrameBase, it.pc, it.regs)
 	return fb
-}
-
-func (it *stackIterator) newStackframe(ret, retaddr uint64) Stackframe {
-	if retaddr == 0 {
-		it.err = NullAddrError{}
-		return Stackframe{}
-	}
-	f, l, fn := it.bi.PCToLine(it.pc)
-	if fn == nil {
-		f = "?"
-		l = -1
-	} else {
-		it.regs.FrameBase = it.frameBase(fn)
-	}
-	r := Stackframe{Current: Location{PC: it.pc, File: f, Line: l, Fn: fn}, Regs: it.regs, Ret: ret, addrret: retaddr, stackHi: it.stackhi, SystemStack: it.systemstack, lastpc: it.pc}
-	r.Call = r.Current
-	if !it.top && r.Current.Fn != nil && it.pc != r.Current.Fn.Entry {
-		// if the return address is the entry point of the function that
-		// contains it then this is some kind of fake return frame (for example
-		// runtime.sigreturn) that didn't actually call the current frame,
-		// attempting to get the location of the CALL instruction would just
-		// obfuscate what's going on, since there is no CALL instruction.
-		switch r.Current.Fn.Name {
-		case "runtime.mstart", "runtime.systemstack_switch":
-			// these frames are inserted by runtime.systemstack and there is no CALL
-			// instruction to look for at pc - 1
-		default:
-			r.lastpc = it.pc - 1
-			r.Call.File, r.Call.Line = r.Current.Fn.cu.lineInfo.PCToLine(r.Current.Fn.Entry, it.pc-1)
-		}
-	}
-	return r
 }
 
 func (it *stackIterator) stacktrace(depth int) ([]Stackframe, error) {
@@ -529,104 +376,6 @@ func (it *stackIterator) appendInlineCalls(frames []Stackframe, frame Stackframe
 	}
 
 	return append(frames, frame)
-}
-
-// advanceRegs calculates it.callFrameRegs using it.regs and the frame
-// descriptor entry for the current stack frame.
-// it.regs.CallFrameCFA is updated.
-func (it *stackIterator) advanceRegs() (callFrameRegs op.DwarfRegisters, ret uint64, retaddr uint64) {
-	fde, err := it.bi.frameEntries.FDEForPC(it.pc)
-	var framectx *frame.FrameContext
-	if _, nofde := err.(*frame.ErrNoFDEForPC); nofde {
-		framectx = it.bi.Arch.FixFrameUnwindContext(nil, it.pc, it.bi)
-	} else {
-		framectx = it.bi.Arch.FixFrameUnwindContext(fde.EstablishFrame(it.pc), it.pc, it.bi)
-	}
-
-	cfareg, err := it.executeFrameRegRule(0, framectx.CFA, 0)
-	if cfareg == nil {
-		it.err = fmt.Errorf("CFA becomes undefined at PC %#x", it.pc)
-		return op.DwarfRegisters{}, 0, 0
-	}
-	it.regs.CFA = int64(cfareg.Uint64Val)
-
-	callimage := it.bi.PCToImage(it.pc)
-
-	callFrameRegs = op.DwarfRegisters{StaticBase: callimage.StaticBase, ByteOrder: it.regs.ByteOrder, PCRegNum: it.regs.PCRegNum, SPRegNum: it.regs.SPRegNum, BPRegNum: it.regs.BPRegNum}
-
-	// According to the standard the compiler should be responsible for emitting
-	// rules for the RSP register so that it can then be used to calculate CFA,
-	// however neither Go nor GCC do this.
-	// In the following line we copy GDB's behaviour by assuming this is
-	// implicit.
-	// See also the comment in dwarf2_frame_default_init in
-	// $GDB_SOURCE/dwarf2-frame.c
-	callFrameRegs.AddReg(uint64(amd64DwarfSPRegNum), cfareg)
-
-	for i, regRule := range framectx.Regs {
-		reg, err := it.executeFrameRegRule(i, regRule, it.regs.CFA)
-		callFrameRegs.AddReg(i, reg)
-		if i == framectx.RetAddrReg {
-			if reg == nil {
-				if err == nil {
-					err = fmt.Errorf("Undefined return address at %#x", it.pc)
-				}
-				it.err = err
-			} else {
-				ret = reg.Uint64Val
-			}
-			retaddr = uint64(it.regs.CFA + regRule.Offset)
-		}
-	}
-
-	return callFrameRegs, ret, retaddr
-}
-
-func (it *stackIterator) executeFrameRegRule(regnum uint64, rule frame.DWRule, cfa int64) (*op.DwarfRegister, error) {
-	switch rule.Rule {
-	default:
-		fallthrough
-	case frame.RuleUndefined:
-		return nil, nil
-	case frame.RuleSameVal:
-		reg := *it.regs.Reg(regnum)
-		return &reg, nil
-	case frame.RuleOffset:
-		return it.readRegisterAt(regnum, uint64(cfa+rule.Offset))
-	case frame.RuleValOffset:
-		return op.DwarfRegisterFromUint64(uint64(cfa + rule.Offset)), nil
-	case frame.RuleRegister:
-		return it.regs.Reg(rule.Reg), nil
-	case frame.RuleExpression:
-		v, _, err := op.ExecuteStackProgram(it.regs, rule.Expression)
-		if err != nil {
-			return nil, err
-		}
-		return it.readRegisterAt(regnum, uint64(v))
-	case frame.RuleValExpression:
-		v, _, err := op.ExecuteStackProgram(it.regs, rule.Expression)
-		if err != nil {
-			return nil, err
-		}
-		return op.DwarfRegisterFromUint64(uint64(v)), nil
-	case frame.RuleArchitectural:
-		return nil, errors.New("architectural frame rules are unsupported")
-	case frame.RuleCFA:
-		if it.regs.Reg(rule.Reg) == nil {
-			return nil, nil
-		}
-		return op.DwarfRegisterFromUint64(uint64(int64(it.regs.Uint64Val(rule.Reg)) + rule.Offset)), nil
-	case frame.RuleFramePointer:
-		curReg := it.regs.Reg(rule.Reg)
-		if curReg == nil {
-			return nil, nil
-		}
-		if curReg.Uint64Val <= uint64(cfa) {
-			return it.readRegisterAt(regnum, curReg.Uint64Val)
-		}
-		newReg := *curReg
-		return &newReg, nil
-	}
 }
 
 func (it *stackIterator) readRegisterAt(regnum uint64, addr uint64) (*op.DwarfRegister, error) {
